Fix class FakingEntityMonkey
Move it closer to tests. Change name to PrepareMocking. Write a few more
lines in docstring
